requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Limiter==3.5.0
Flask-Babel==4.0.0
Werkzeug==2.3.7
Pillow==10.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
